Seperate locationmanager from mypathview
